Creates utils.step_reader module
